Complete IdleKV codebase review and fix for workshop paper
🎯 ALL 6 TASKS COMPLETED - Ready for SCALE @ ICML 2026 deadline

## Task 1: Fixed All 6 Critical Bugs ✅
- Bug 1: DynamicCache format compatibility via new kv_cache.py utilities
- Bug 2: RoPE application in phase1_rescore.py with pragmatic position-0 fix
- Bug 3: O(n²) memory fix in compression.py using list accumulation
- Bug 4: Hidden states capture via model.config.output_hidden_states = True
- Bug 5: Thread safety in scheduler.py using threading.Event
- Bug 6: GPU-optimized FullKVStore (96GB VRAM) with CPU fallback option

## Task 2: All Tests Pass on CPU (Mac) ✅
- 20/20 tests passing with comprehensive coverage
- New test_kv_cache.py for DynamicCache/tuple compatibility
- New test_compression.py for end-to-end CompressedKVManager
- Fixed Python 3.9 compatibility issues
- All tests use device='cpu' and dtype=float32 for Mac

## Task 3: Go/No-Go Script Runnable ✅
- Device handling with auto-detection and graceful CPU fallback
- Fixed position tracking using actual compressed cache size
- Manual greedy decoding replacing model.generate() complexity
- Proper warmup loop populating query buffer and tracking KV
- Added --device flag and comprehensive error handling

## Task 4: Experiment Runner Complete ✅
- Full framework generating 99 experiments (baselines + IdleKV + ablations)
- RULER: Needle-in-a-haystack evaluation (1-3 needles)
- LongBench: F1-scored QA with dataset loading
- Simulation harness: Tool-call pause simulation
- kvpress baseline handling with mock fallback
- Structured JSON results format

## Task 5: Professional README ✅
- Clear setup: Mac development + GPU experiments
- Step-by-step workflow: go/no-go → experiments → figures
- Complete repository structure tree
- Hardware notes: 96GB VRAM optimization + smaller GPU fallback
- Zero-to-running in <10 minutes (excluding model download)

## Task 6: Hardware-Optimized Configs ✅
- Added hardware: rtx_pro_6000_blackwell_96gb
- Set offload_full_kv: false for 96GB GPU setup
- Created configs/models.yaml with Llama-3.1-8B + Qwen2.5-7B specs
- Verified kvpress class name compatibility

## Verification: ALL REQUIREMENTS MET ✅
✅ Mac (CPU): pip install -e .[dev] && pytest tests/ -v (20/20 pass)
✅ Dry run: scripts/run_experiments.py --dry-run (99 experiments)
✅ Ready for GPU: scripts/go_no_go.py + full experiment suite
✅ Professional documentation and robust error handling
✅ 96GB VRAM optimizations + smaller GPU graceful fallback

🚀 RESEARCH-READY: Complete experimental framework for workshop paper evaluation

+++ tests/test_shadow_buffer.py
@@ -6,7 +6,7 @@ from idlekv.core.shadow_buffer import ShadowBuffer
 
 
 def test_push_and_get():
-    buf = ShadowBuffer(num_layers=2, max_size=4, num_kv_heads=2, head_dim=8, device="cpu")
+    buf = ShadowBuffer(num_layers=2, max_size=4, num_kv_heads=2, head_dim=8, device="cpu", dtype=torch.float32)
 
     k = torch.randn(2, 3, 8)  # 3 evicted tokens
     v = torch.randn(2, 3, 8)
@@ -57,7 +57,7 @@ def test_clear():
 
 def test_overflow():
     """Push more tokens than max_size at once."""
-    buf = ShadowBuffer(num_layers=1, max_size=4, num_kv_heads=1, head_dim=4, device="cpu")
+    buf = ShadowBuffer(num_layers=1, max_size=4, num_kv_heads=1, head_dim=4, device="cpu", dtype=torch.float32)
     k = torch.randn(1, 10, 4)  # 10 > max_size=4
     v = torch.randn(1, 10, 4)
     buf.push(0, k, v)
